Fix the handling of file paths
Some tasks used the File(String) constructor to attempt to open files in
the project directory. Gradle does not have to (and it doesn't) use the
project directory as the working directory. As a result, these tasks
were broken. Switch to the correct way of creating Files under Gradle,
Project#file.