Fix non-main source set remap configurations publishing by default (#891)
Fixes #890.

Note that there is *no simple way* to enable publishing for remap
configurations created this way anymore as the publishing mode is read
during the `createRemapConfigurations` call.

This also changes the behaviour of `modClientX` configurations to no longer
publish. If that is wanted, I can add a `createRemapConfigurations()`
overload with some kind of configurability.